python-ml/requirements.txt
numpy==1.18.2
pandas==1.0.3
pymongo==3.10.1
scikit-learn==0.22.2.post1
pydash==4.7.6
python-dotenv==0.12.0